requirements.txt
pillow==7.2.0
backports.zoneinfo==0.2.1
requests-mock==1.8.0
requests==2.24.0
wrapt==1.12.1
